# swagger.yaml
swagger: "2.0"
info:
  description: "Vote Events"
  version: "1.0.0"
  title: "Event shuffle"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/api/v1"
...
schemes:
 - "http"
paths:
...
  /event/{eventId}/vote:
    post:
      tags:
      - "event"
      summary: "Add a new vote to event"
      description: ""
      operationId: "addVote"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Vote object that needs to be added to the db"
        required: true
        schema:
          $ref: "#/definitions/Vote"
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      responses:
        "405":
          description: "Invalid input"
...
definitions:
...
  Vote:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name of the Voter"
      date:
        type: "array"
        xml:
          name: "dates"
          wrapped: true
        items:
          type: "string"
    xml:
      name: "Vote"
...